# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amrutam Telemedicine API
  version: 1.0.0
  description: |
    Production-grade REST API for an Ayurveda telemedicine platform.

    ## Authentication
    Protected endpoints require a JWT bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.

    ## Idempotency
    `POST /api/v1/appointments` requires an `Idempotency-Key` header (UUID recommended).

    ## Response Envelope
    Success: `{ success, data, meta, requestId, timestamp }`
    Error: `{ success: false, code, message, details, requestId, timestamp }`
  contact:
    name: Amrutam Engineering
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Client-generated UUID for idempotent booking
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, code, message, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    CreateAppointmentRequest:
      type: object
      required: [doctorId, slotId]
      properties:
        doctorId:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 500
        notes:
          type: string
          maxLength: 1000
    RescheduleAppointmentRequest:
      type: object
      required: [newSlotId]
      properties:
        newSlotId:
          type: string
          format: uuid
        reason:
          type: string
    CancelAppointmentRequest:
      type: object
      properties:
        reason:
          type: string
...
  responses:
    Unauthorized:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: UNAUTHORIZED
            message: Invalid or missing bearer token
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    Forbidden:
      description: Insufficient role or resource access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Insufficient role or resource access
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: VALIDATION_ERROR
            message: Request validation failed
            details: ['doctorId must be a UUID']
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: NOT_FOUND
            message: Resource not found
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Rate limit exceeded
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: INTERNAL_ERROR
            message: An unexpected error occurred
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
paths:
...
  /appointments:
    post:
      tags: [Appointments]
      summary: Book an appointment
      description: |
        Creates appointment with idempotent semantics inside a single DB transaction:
        optimistic slot locking, audit log, and transactional outbox event.
        Requires Patient role and `Idempotency-Key` header.
      security:
        - bearerAuth: []
        - idempotencyKey: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            example:
              doctorId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              slotId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
              reason: Digestive issues
      responses:
        '201':
          description: Appointment booked
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: CONFIRMED
                  scheduledStart: '2026-07-15T10:00:00.000Z'
                  scheduledEnd: '2026-07-15T10:30:00.000Z'
                  doctor:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    consultationFee: 500
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a patient, inactive account, or doctor self-booking
          content:
            application/json:
              examples:
                notPatient:
                  value:
                    success: false
                    code: FORBIDDEN
                    message: Only patients can book appointments
                selfBooking:
                  value:
                    success: false
                    code: DOCTOR_SELF_BOOKING
                    message: Doctors cannot book their own slots
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Slot conflict or idempotency conflict
          content:
            application/json:
              examples:
                slotBooked:
                  value:
                    success: false
                    code: SLOT_ALREADY_BOOKED
                    message: Slot is no longer available
                idempotencyConflict:
                  value:
                    success: false
                    code: IDEMPOTENCY_CONFLICT
                    message: Idempotency key reused with different request payload
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /appointments/me:
    get:
      tags: [Appointments]
      summary: List patient appointments
      description: Cursor-paginated list for authenticated patient. Requires Patient role.
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, NO_SHOW]
      responses:
        '200':
          description: Paginated appointments
          content:
            application/json:
              example:
                success: true
                data:
                  items: []
                  nextCursor: null
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /appointments/doctor/me:
    get:
      tags: [Appointments]
      summary: List doctor appointments
      description: Cursor-paginated list for authenticated doctor. Requires Doctor role.
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated appointments
          content:
            application/json:
              example:
                success: true
                data:
                  items: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by ID
      description: Returns appointment if caller is patient, assigned doctor, or admin.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: CONFIRMED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /appointments/{id}/cancel:
    patch:
      tags: [Appointments]
      summary: Cancel an appointment
      description: Validates status transition, releases slot, writes audit and outbox event. Patient, doctor, or admin may cancel per service rules.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentRequest'
            example:
              reason: Schedule conflict
      responses:
        '200':
          description: Appointment cancelled
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: CANCELLED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cancellation not allowed
          content:
            application/json:
              example:
                success: false
                code: CANCELLATION_NOT_ALLOWED
                message: Appointment cannot be cancelled in current status
        '500':
          $ref: '#/components/responses/InternalError'
  /appointments/{id}/reschedule:
    patch:
      tags: [Appointments]
      summary: Reschedule appointment to new slot
      description: Reserves new slot with optimistic lock, releases old slot, records history. Patient role required.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleAppointmentRequest'
            example:
              newSlotId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
              reason: Prefer afternoon slot
      responses:
        '200':
          description: Appointment rescheduled
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: CONFIRMED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Slot unavailable or reschedule not allowed
          content:
            application/json:
              examples:
                slotBooked:
                  value:
                    success: false
                    code: SLOT_ALREADY_BOOKED
                    message: Slot is no longer available
                notAllowed:
                  value:
                    success: false
                    code: RESCHEDULE_NOT_ALLOWED
                    message: Appointment cannot be rescheduled
        '500':
          $ref: '#/components/responses/InternalError'
...